add slack notifications
