+++ etl/metadata_export.py
@@ -5,6 +5,7 @@ from typing import Any, Dict, List, Literal, Optional, Tuple, Union
 import pandas as pd
 import rich_click as click
 from owid.catalog import Dataset, utils
+from owid.catalog.yaml_metadata import _merge_variable_metadata
 from rich.console import Console
 from rich.syntax import Syntax
 
@@ -91,14 +92,24 @@ def merge_or_create_yaml(meta_dict: Dict[str, Any], output_path: Path) -> str:
 
         for k, v in meta_dict.get("dataset", {}).items():
             doc["dataset"].setdefault(k, v)
+
         for tab_name, tab_dict in meta_dict.get("tables", {}).items():
             variables = tab_dict.pop("variables", {})
+
+            # update table
             if tab_name not in doc["tables"]:
                 doc["tables"][tab_name] = {}
             doc["tables"][tab_name].update(tab_dict)
+
+            # update variables
             if "variables" not in doc["tables"][tab_name]:
                 doc["tables"][tab_name]["variables"] = {}
-            doc["tables"][tab_name]["variables"].update(variables)
+            orig_variables = doc["tables"][tab_name]["variables"]
+
+            for var_name, var_meta in variables.items():
+                orig_variables[var_name] = _merge_variable_metadata(
+                    orig_variables[var_name], var_meta, if_origins_exist="replace", overwrite=True
+                )
 
         if doc["dataset"] == {}:
             del doc["dataset"]
